refactor: 뒤로가기 버튼 children으로 받도록 수정
