+++ src/components/common/PageTitle.js
@@ -1,22 +1,9 @@
-import { useNavigate } from 'react-router-dom';
+import { useNavigate } from "react-router-dom";
 
-import styled from 'styled-components';
+import styled from "styled-components";
 
-export const PageTitle = ({ children, isRoot }) => {
-  const navigate = useNavigate();
-
-  return (
-    <PageTitleWrapperStyle>
-      {!isRoot && (
-        <BackwardButton
-          onClick={() => {
-            navigate(-1);
-          }}
-        />
-      )}
-      {children}
-    </PageTitleWrapperStyle>
-  );
+export const PageTitle = ({ children }) => {
+  return <PageTitleWrapperStyle>{children}</PageTitleWrapperStyle>;
 };
 
 const PageTitleWrapperStyle = styled.div`
@@ -25,13 +12,4 @@ const PageTitleWrapperStyle = styled.div`
   display: flex;
   justify-content: space-between;
   align-items: center;
-`;
-
-const BackwardButton = styled.span`
-  width: 8px;
-  height: 8px;
-  border: 1px;
-  border-style: none none solid solid;
-  transform: rotate(45deg);
-  cursor: pointer;
 `;
